fix: :bug: Make error state recoverable by transitioning back to idle + inline machine config
